requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.3.0
python-dotenv==0.21.0
pytube==12.1.0
qrcode==7.3.1
Werkzeug==2.2.2
